We (B.E.T) have a loop now
We created a for loop to reduce the repetition in this script.  We defined the variable as N, a population of 10 individuals.  To repeat the same simulation ten times, we made a for loop.  The for loop calls the simulation to be run 10 times (i in 1:10). The for loop also contains an 'if, else' statement that would allow the data to be plotted without negative populations. Then we set our parameters in the form of vital rates, with equal birth and death rates. (Bryce, Erica, Tomoe)